Create clone section in install script
This commit creates a new section in the install script that clones the
project to the users local machine as well as setting up the git
credential.helper store to remove the need from providing the dhbw
gitlab credentials over and over again.

This cloning also sets up the remotes in the created local repository,
mainly upstream pointing to the latex-template repository and origin to
the provided repository.