Check for git name and email before adding modules
As modules are created and committed to the git repository the project
lives in, git requires to be configured for user.name and user.email.
If those two values do not exist, git cannot commit the changes created
by installing a module and therefore fails.

This commit adds a simply check prior to module installation that checks
the configuration and asks for the not yet set values.